2009-04-27  Vladimir Serbinenko  <phcoder@gmail.com>
	Warning fix

	* disk/scsi.c (grub_scsi_open): added missing cast when 
	calling grub_dprintf

+++ disk/scsi.c
@@ -301,7 +301,8 @@ grub_scsi_open (const char *name, grub_disk_t disk)
 				 << GRUB_DISK_SECTOR_BITS);
 
 	  grub_dprintf ("scsi", "capacity=%llu, blksize=%d\n",
-			disk->total_sectors, scsi->blocksize);
+			(unsigned long long) disk->total_sectors, 
+			scsi->blocksize);
 
 	  return GRUB_ERR_NONE;
 	}
